BREAKING: return image's width and height
Instead of returning the image as a Buffer, this package now returns an object with three properties: "image", "width", and "height". The "image" property is the same as what was returned before.

To upgrade, anywhere you have this:

const buf = captureFrame(video, format)

Change it to this:

const buf = captureFrame(video, format)?.image